requirements.txt
ttp==0.8.0
openpyxl==3.0.9
click==8.0.1
netmiko==3.4.0
pandas==1.3.3
PyYAML==5.4.1
